+++ R/00_global_vars.R
@@ -9,4 +9,4 @@ globalVariables(c("index", "index.num", "y", "yhat", "week", "mday", "diff.media
                   ".metric", ".metrics", ".rank_metric", ".rank_std_err", "failure_rate", "std_err",
                   ".rank_failure_rate", ".txt", "training", "testing", "fitted", "model", "..val", "..nm",
                   ".white_noise_lower", ".white_noise_upper", ".date_var_collapsed", ".rowid", "..date_agg", "grp_names",
-                  "check_row_exists"))
+                  "check_row_exists", ".box_group"))