+++ NEWS.md
@@ -4,7 +4,7 @@ __New Features__
 
 - `tk_tsfeatures()`: A new function that makes it easy to generate time series feature matrix using `tsfeatures`. The main benefit is that you can pipe time series data in `tibbles` with `dplyr` groups. The features will be produced by group. #95 #84
 
-- 
+- `plot_time_series_boxplot()`: A new function that makes plotting time series boxplots simple using a `.period` argument for time series aggregation. 
 
 __New Vignettes__