+++ NEWS.md
@@ -1,4 +1,4 @@
-# timetk 2.6.2.9000
+# timetk 2.7.0
 
 __New Features__